model-training/reid/requirements.txt
# ReID Module Requirements

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# YOLO detector
ultralytics>=8.0.0

# Data processing
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Training utilities
tensorboard>=2.14.0
pyyaml>=6.0.0
tqdm>=4.66.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
